requirements.txt
# Core Dependencies
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
lxml>=4.9.0
beautifulsoup4>=4.12.0

# NLP and Text Processing
nltk>=3.8.0
spacy>=3.6.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0
networkx>=3.1.0

# Fuzzy Logic
scikit-fuzzy>=0.4.2

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wordcloud>=1.9.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0

# Jupyter and Documentation
jupyter>=1.0.0
notebook>=6.5.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Optional: GPU acceleration
# torch-audio>=2.0.0  # Uncomment if needed
# torch-vision>=0.15.0  # Uncomment if needed
